Created child component timer_controls.  Changed pomodoro timer to implement timer controls.  Removed unused variable from pomodoro-timer.
